Folder refactor and fixed bugs in code
