+++ src/app/feature-libs/organization/ey-user-registration/components/form/ey-success-page/ey-success-page.component.spec.ts
