+++ src/app/feature-libs/organization/ey-user-registration/components/form/ey-success-page/ey-success-page.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { UrlModule, I18nModule } from '@spartacus/core';
+import { EySuccessPageComponent } from './ey-success-page.component';
+
+@NgModule({
+  declarations: [EySuccessPageComponent],
+  imports: [CommonModule, UrlModule, I18nModule],
+})
+export class EySuccessPageModule {}
