+++ src/app/feature-libs/organization/ey-user-registration/components/form/ey-user-registration-form.component.scss
