+++ src/app/feature-libs/organization/ey-user-registration/components/form/ey-user-registration-form.service.ts
@@ -39,6 +39,7 @@ export class EyUserRegistrationFormService {
       titleCode: [null],
       firstName: ['', Validators.required],
       lastName: ['', Validators.required],
+      gender: [null, Validators.required],
       companyName: ['', Validators.required],
       email: ['', [Validators.required, CustomFormValidators.emailValidator]],
       country: this.formBuilder.group({
@@ -180,17 +181,17 @@ export class EyUserRegistrationFormService {
           firstName: form.get('firstName')?.value,
           lastName: form.get('lastName')?.value,
           email: form.get('email')?.value,
-          gender: form.get('gender')?.value,
-          dob: form.get('dob')?.value,
-          age: form.get('age')?.value,
-          qualification: form.get('qualification')?.value,
-          companyName: form.get('companyName')?.value,
-          isocode: form.get('isocode')?.value,
-          line1: form.get('line1')?.value,
-          line2: form.get('line2')?.value,
-          town: form.get('town')?.value,
-          postalCode: form.get('postalCode')?.value,
-          phoneNumber: form.get('phoneNumber')?.value,
+          // gender: form.get('gender')?.value,
+          // dob: form.get('dob')?.value,
+          // age: form.get('age')?.value,
+          // qualification: form.get('qualification')?.value,
+          // companyName: form.get('companyName')?.value,
+          // isocode: form.get('isocode')?.value,
+          // line1: form.get('line1')?.value,
+          // line2: form.get('line2')?.value,
+          // town: form.get('town')?.value,
+          // postalCode: form.get('postalCode')?.value,
+          // phoneNumber: form.get('phoneNumber')?.value,
           message: message,
         })
       ),
