+++ src/app/feature-libs/user/account/assets/ng-package.json
