+++ src/app/feature-libs/user/account/components/ey-login-form/ey-login-form.module.ts
@@ -22,6 +22,7 @@ import { LoginFormComponentService } from '@spartacus/user/account/components';
 import { EyLoginFormComponent } from './ey-login-form.component';
 
 @NgModule({
+  declarations: [EyLoginFormComponent],
   imports: [
     CommonModule,
     FormsModule,
