+++ src/app/spartacus/features/organization/organization-user-registration-wrapper.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CDCB2BRegisterModule } from '@spartacus/cdc/organization/user-registration';
-import { EyOrganizationUserRegistrationModule } from '../../../../feature-libs/organization/ey-user-registration/ey-user-registration.module';
+import { EyOrganizationUserRegistrationModule } from '../../../feature-libs/organization/ey-user-registration/ey-user-registration.module';
 
 @NgModule({
   declarations: [],
