+++ src/app/spartacus/features/user/user-account-wrapper.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CDCUserAccountModule } from '@spartacus/cdc/user-account';
-import { UserAccountModule } from '../../../../feature-libs/user/account/user-account.module';
+import { UserAccountModule } from '../../../feature-libs/user/account/user-account.module';
 
 @NgModule({
   declarations: [],
