+++ src/app/spartacus/spartacus-features.module.ts
@@ -84,7 +84,7 @@ import { ProductVariantsFeatureModule } from './features/product/product-variant
 import { ProductImageZoomFeatureModule } from './features/product/product-image-zoom-feature.module';
 import { QuoteFeatureModule } from './features/quote/quote-feature.module';
 import { OrganizationUserRegistrationFeatureModule } from './features/registration-feature.module';
-import { EySuccessPageModule } from '../../feature-libs/organization/ey-user-registration/components/form/ey-success-page/ey-success-page.module';
+import { EySuccessPageModule } from '../feature-libs/organization/ey-user-registration/components/form/ey-success-page/ey-success-page.module';
 
 @NgModule({
   declarations: [],
