--- src/feature-libs/organization/ey-user-registration/components/form/ey-success-page/ey-success-page.module.ts
@@ -1,39 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { RouterModule } from '@angular/router';
-import {
-  UrlModule,
-  I18nModule,
-  FeaturesConfigModule,
-  ConfigModule,
-  NotAuthGuard,
-  CmsConfig,
-} from '@spartacus/core';
-import {
-  FormErrorsModule,
-  SpinnerModule,
-  PasswordVisibilityToggleModule,
-} from '@spartacus/storefront';
-import { EySuccessPageComponent } from './ey-success-page.component';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule,
-    UrlModule,
-    I18nModule,
-    FormErrorsModule,
-    SpinnerModule,
-    PasswordVisibilityToggleModule,
-    FeaturesConfigModule,
-    ConfigModule.withConfig({
-      cmsComponents: {
-        EySuccessPageComponent: {
-          component: EySuccessPageComponent,
-          guards: [NotAuthGuard],
-        },
-      },
-    } as CmsConfig),
-  ],
-})
-export class EySuccessPageModule {}
